Use BPF_F_REUSE_STACKID
Not nice to override - but as far as I understand there's no better, simple option;
unless we reuse the same buckets, this map will continue growing forever.
Clearing it from time to time is not an optino either - things are being added, concurrently.
Need to think of a more complex solution (like swap between 2 maps, clean, swap again ....).

For now, this is certainly good enough.